Session Class moved to Session Factory
