+++ csm/conf/csm_default_db.conf
@@ -43,7 +43,7 @@ models:
     config:
       consul_db:
         collection: replace_node_data
-  - import_path: csm.core.services.sessions.Session
+  - import_path: csm.core.services.session_factory.Session
     database: consul_db
     config:
         consul_db:
