+++ csm/conf/etc/csm/database.conf
@@ -50,7 +50,7 @@ models:
     config:
         consul_db:
             collection: "replace_node_data"
--   import_path: "csm.core.services.sessions.Session"
+-   import_path: "csm.core.services.session_factory.Session"
     database: "consul_db"
     config:
         consul_db:
