+++ csm/core/services/session_factory.py
@@ -13,12 +13,65 @@
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 
-from csm.core.services.sessions import Session
 from typing import Optional
 from cortx.utils.data.db.db_provider import DataBaseProvider
 from csm.core.blogic import const
 from cortx.utils.data.access import Query
 from cortx.utils.data.access.filters import Compare
+from csm.core.services.permissions import PermissionSet
+
+class SessionCredentials:
+    """ Base class for a variying part of the session
+    depending on the user type (CSM, LDAP, S3).
+    """
+
+    def __init__(self, user_id: str) -> None:
+        self._user_id = user_id
+
+    @property
+    def user_id(self) -> str:
+        return self._user_id
+
+
+class Session(CsmModel):
+    """ Session data """
+
+    Id = str
+    _id = "_session_id"
+
+    def __init__(self, session_id: Id,
+                 expiry_time: datetime,
+                 credentials: SessionCredentials,
+                 permissions: PermissionSet) -> None:
+        self._session_id = session_id
+        self._expiry_time = expiry_time
+        self._credentials = credentials
+        self._permissions = permissions
+
+    @property
+    def session_id(self) -> Id:
+        return self._session_id
+
+    @property
+    def expiry_time(self) -> datetime:
+        return self._expiry_time
+
+    @expiry_time.setter
+    def expiry_time(self, expiry_time):
+        self._expiry_time = expiry_time
+
+    @property
+    def credentials(self) -> SessionCredentials:
+        return self._credentials
+
+    @property
+    def permissions(self) -> PermissionSet:
+        return self._permissions
+
+    def get_user_role(self) -> Optional[str]:
+        creds = self._credentials
+        return creds.user_role if isinstance(creds, LocalCredentials) else None
+
 
 class InMemory:
     def __init__(self):
@@ -64,3 +117,4 @@ class SessionFactory:
             return InMemory()
         else:
             raise ValueError(session_backend)
+
